+++ packages/core/src/constants.ts
@@ -132,6 +132,9 @@ export const AI_MODELS: ModelOption[] = [
 
 export const DEFAULT_AI_MODEL = AI_MODELS[0].id
 
+export const AUTOMATION_HTTP_PORT = 7600
+export const AUTOMATION_WS_PORT = 7601
+
 export const GOOGLE_FONTS_API_KEY = 'AIzaSyD1tYDR_dUEiV-Tw1vksEhZbUytgKW5pc8'
 
 export const CJK_FALLBACK_FAMILIES_MACOS = [